tests/wpt/web-platform-tests/css/css-scroll-snap/parsing/scroll-padding-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-padding with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-padding">
<meta name="assert" content="scroll-padding supports only the grammar '[ <length-percentage> | auto ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("scroll-padding-top", "20");
test_invalid_value("scroll-padding-top", "-20px");


test_invalid_value("scroll-padding-right", "20");
test_invalid_value("scroll-padding-right", "-20px");


test_invalid_value("scroll-padding-bottom", "20");
test_invalid_value("scroll-padding-bottom", "-20px");


test_invalid_value("scroll-padding-left", "20");
test_invalid_value("scroll-padding-left", "-20px");


test_invalid_value("scroll-padding", "20");
test_invalid_value("scroll-padding", "-20px");
</script>
</body>
</html>
